Readability and improvements for future implementations
